refactor(session): extract stateless helpers to session_support module
session.py was a 5000-line god object that PR #407 deliberately left
untouched. This is the first of a stacked series that converges it toward
a coordinator role by moving genuinely separable pieces out.

PR1 is the zero-behaviour-risk slice: the module-level helpers that never
touched ``self`` and several of which already leaked out as a de-facto
public API (imported by cli/tools, diagnostic/startup, and the test suite).
Moved to a new ``loom/core/session_support.py``:

  - envelope metadata markers (_OUTCOME_MARKERS, _clean_envelope_metadata_line,
    _iter_nonempty_lines, _extract_envelope_intent/outcome_marker)
  - think-block streaming lookahead (_find_partial_tag_suffix)
  - output-token resolution (_resolve_output_max_tokens + constants)
  - .env loading (_load_env)
  - skill frontmatter parsing (_parse_skill_frontmatter)
  - background-jobs reminder (_build_jobs_inject_message)
  - scope confirm-prompt formatting (_format_scope_panel,
    _format_scope_summary_plain — were @staticmethod, now module functions)

session.py re-exports all of them, so historical import paths
(``from loom.core.session import _load_env`` etc.) keep working untouched.
The two scope formatters were the only behaviour-visible move: their one
test reached them via ``LoomSession._format_scope_panel``; updated to the
re-exported module function.

session.py: 5000 → 4665 lines. Full suite green (2252 passed; the 3
pre-existing failures — version-sync + a discord mock — are unrelated and
fail identically on master).
